convert projects to .net sdk format and modify design of rocksdbnative to deploy an assembly to act as a beacon
This fixes the problem with using RocksDbSharp in a unit test environment as well as any environment that uses shadow copying or runs from a different codebase than usual.

The native libraries have been removed from the RocksDbSharp csproj and moved to a new RocksDbNative csproj.  The RocksDbNative csproj now also builds an assembly with a single class and a single method for getting the codebase.  If you previously had a project reference to the RocksDbSharp project, native binaries were automatically copied for you.  This is no longer true and not desired.  Now the project reference works like the package reference and you should add a project reference to RocksDbNative to get the native binaries.